refactor: Create a generic questions form
